+++ src/components/data-display/tables/DomainsTable.tsx
@@ -106,7 +106,8 @@ function filterTableData(filter: string, data: TableData[]): TableData[] {
 const DomainsTable = ({
   domainData,
   loading,
-  filter,
+  filter = '',
+  setFilter,
 }: {
   domainData: {
     names: Record<string, AoArNSNameData>;
@@ -114,6 +115,7 @@ const DomainsTable = ({
   };
   loading: boolean;
   filter?: string;
+  setFilter: (filter: string) => void;
 }) => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
@@ -202,11 +204,8 @@ const DomainsTable = ({
   ]);
 
   useEffect(() => {
-    if (filter) {
-      setFilteredTableData(filterTableData(filter, tableData));
-    } else {
-      setFilteredTableData([]);
-    }
+    const filtered = filterTableData(filter, tableData);
+    setFilteredTableData(filtered);
   }, [filter, tableData]);
   // Define columns for the table
   const columns: ColumnDef<TableData, any>[] = [
@@ -554,14 +553,14 @@ const DomainsTable = ({
       <div className="w-full">
         <TableView
           columns={columns}
-          data={
-            filteredTableData.length
-              ? filteredTableData
-              : tableData.length
-              ? tableData
-              : []
-          }
+          data={filteredTableData}
           isLoading={false}
+          onRowClick={(rowData, tableRow) => {
+            if (tableRow) {
+              tableRow.toggleExpanded();
+            }
+            return rowData;
+          }}
           noDataFoundText={
             !walletAddress ? (
               <div className="flex flex-column text-medium center white p-[100px] box-border gap-[20px]">
@@ -581,6 +580,21 @@ const DomainsTable = ({
             ) : loading ? (
               <div className="flex flex-column center white p-[100px]">
                 <Loader message="Loading assets..." />
+              </div>
+            ) : // if a filter is provided, show the no data found message
+            filter && filter.length > 0 ? (
+              <div className="flex flex-column center p-[100px]">
+                <span className="white bold" style={{ fontSize: '16px' }}>
+                  No results found for &apos;{filter}&apos;
+                </span>
+                <button
+                  onClick={() => {
+                    setFilter('');
+                  }}
+                  className="button-secondary center p-[10px] w-fit"
+                >
+                  Clear filter
+                </button>
               </div>
             ) : (
               <div className="flex flex-column center p-[100px]">
@@ -608,7 +622,7 @@ const DomainsTable = ({
           }
           defaultSortingState={{
             id: sortBy,
-            desc: sortBy == 'expiryDate' ? false : true,
+            desc: false, // sort by ascending by default
           }}
           renderSubComponent={({ row }) => (
             <UndernamesSubtable