+++ CHANGELOG.md
@@ -8,6 +8,36 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.15.4] - 2025-08-04
+
+### Fixed
+
+- Default to arweave.net for Arweave.app and Metamask wallets
+- Fix Turbo Credits top up modal so app doesn't crash.
+
+### Changed
+
+- Remove wallet-based gateway config and only use app-based gateway config
+
+## [1.15.3] - 2025-07-31
+
+### Changed
+
+- Open undername table on row click in domains table.
+
+## [1.15.2] - 2025-07-31
+
+### Changed
+
+- Fixed filter functionality in domains table to show no results found message
+  when no results are found.
+
+## [1.15.1] - 2025-07-31
+
+### Changed
+
+- Changed default sort order to ascending for domains table.
+
 ## [1.15.0] - 2025-07-30
 
 ### Added